+++ src/libwdb/ars.c
@@ -49,7 +49,7 @@
  * -1 Fail
  */
 int
-mk_ars(struct rt_wdb *filep, const char *name, int ncurves, int pts_per_curve, fastf_t **curves)
+mk_ars(struct rt_wdb *filep, const char *name, size_t ncurves, size_t pts_per_curve, fastf_t **curves)
 {
     struct rt_ars_internal *ars;
